+++ scripts/src/react/components/formField/formField.spec.js
@@ -1,0 +1,6 @@
+/**
+ * @todo add tests
+ */
+test('form field', () => {
+  expect(1).toEqual(1);
+});