+++ CHANGELOG.md
@@ -1,7 +1,19 @@
+## [1.0.4] - 2017-07-17 - Patch
+
+### Added
+- Form Field react component
+- Webpack and demo http server for previews
+
+## [1.0.3] - 2017-07-12 - Patch
+
+### Fixed
+- Fixed versioning issues
+
 ## [1.0.2] - 2017-07-12 - Patch
 
 ### Added
 - Adding peer depedency installation script
+- More unit tests
 
 ## [1.0.1] - 2017-07-11 - Patch
 
@@ -26,4 +38,6 @@
 
 [1.0.0]: https://bitbucket.org/byndops/bynd-frontend/commits/b966adb9f1f756212347042e89213c484aa447a4
 [1.0.1]: https://bitbucket.org/byndops/bynd-frontend/commits/e7d399c2e32e47083a463263935e4c27294dee8a
-[1.0.2]: https://bitbucket.org/byndops/bynd-frontend/commits/e7d399c2e32e47083a463263935e4c27294dee8a
+[1.0.2]: https://bitbucket.org/byndops/bynd-frontend/commits/a9db8bc2650d1dda4342f2faca4f632df804d33e
+[1.0.3]: https://bitbucket.org/byndops/bynd-frontend/commits/0f0b6d8a48bb4032968a2e6c68a5e15125cca2d0
+[1.0.4]: https://bitbucket.org/byndops/bynd-frontend/commits/cfe1ab2d99709e99089263f9e06acf221efb4231